Shields no longer engraveable w/ armor engraving tool.
