+++ Scripts/Scripts.csproj
@@ -694,7 +694,6 @@
     <Compile Include="Items\StoreBought\SpiralStaircase.cs" />
     <Compile Include="Items\StoreBought\SquirrelMailbox.cs" />
     <Compile Include="Items\StoreBought\WineRack.cs" />
-    <Compile Include="Items\Tools\ShieldEngravingTool.cs" />
     <Compile Include="Items\Tools\SteamPoweredBeverageMaker.cs" />
     <Compile Include="Items\Addons\ClawFootTub.cs" />
     <Compile Include="Items\Addons\GreenMarbleFireplace.cs" />
